requirement.txt

tqdm
pillow
torch
tensorboard
torchvision
tensorboardX
scikit-image
compressai
opencv-python
numpy